requirements.txt
matplotlib==3.3.2
numpy==1.19.5
pandas==1.3.5
pgmpy==0.1.17
Pillow==9.2.0
scikit_learn==1.1.2
scipy==1.6.2
sympy==1.10.1
torch==1.11.0
opencv-python==4.6.0.66